Moved KeyPresentPage to KeyPresent module
This makes responsibilities more clear. Having that widget in the
SignPages module is not as discoverable.
